Added complete English wikipedia benchmark.
This includes full page text, not just abstracts. The XML dump was first
expanded from the Wiki syntax into HTML so that Omega could better index
it, then indexed into eight separate DBs. A stub database file automates
searching over all separate DBs and merging of results.

The queries are the same 50 as the abstracts, but I've added an expanded
version of them that uses a full sentence.